menus/domain/* -> domain/menus/*
