+++ src/main/kotlin/com/kitchenforce/domain/menus/Menu.kt
@@ -1,4 +1,4 @@
-package com.kitchenforce.menus.domain
+package com.kitchenforce.domain.menus
 
 import org.springframework.data.annotation.CreatedDate
 import org.springframework.data.jpa.domain.support.AuditingEntityListener
@@ -8,7 +8,7 @@ import javax.persistence.*
 @Entity
 @Table
 @EntityListeners(AuditingEntityListener::class)
-class MenuProduct(
+class Menu(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Int?,
@@ -18,12 +18,15 @@ class MenuProduct(
     val createdAt: Date,
 
     @Column(nullable = false)
-    val quantity: Int,
+    val name: String,
+
+    @Column(nullable = false)
+    val price: Int,
+
+    @Column(nullable = false)
+    val isHidden: Boolean,
 
     @ManyToOne
-    @JoinColumn(name = "menu_id")
-    val menu: Menu,
-
-    // TODO: Product 구현시 추가 예정
-//    val product: Product,
+    @JoinColumn(name = "menu_group_id")
+    val menuGroup: MenuGroup,
 )
